# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: FADESP - Pagamentos API
  version: "1.0.0"
  description: |
    API para gerenciamento de pagamentos (criação, consulta, atualização de status e exclusão lógica).
servers:
  - url: http://localhost:8080
    description: Local

tags:
  - name: Pagamentos
    description: Endpoints para criação, consulta, atualização e exclusão lógica de pagamentos

paths:
  /api/pagamentos:
    post:
      tags: [Pagamentos]
      summary: Criar pagamento
      description: Registra um novo pagamento no sistema (status inicial **PENDENTE**).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoRequest'
            examples:
              exemploPix:
                summary: Exemplo com PIX
                value:
                  codigoDebito: 123456
                  cpfCnpj: "12345678901"
                  metodoPagamento: "PIX"
                  valor: 250.00
              exemploCartao:
                summary: Exemplo com Cartão de Crédito
                value:
                  codigoDebito: 987654
                  cpfCnpj: "12345678901234"
                  metodoPagamento: "CARTAO_CREDITO"
                  numeroCartao: "5555444433331111"
                  valor: 99.90
      responses:
        '201':
          description: Pagamento criado com sucesso
          headers:
            Location:
              description: URL do recurso criado
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagamentoResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Conflito (código de débito já utilizado)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          description: Erro interno
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
    get:
      tags: [Pagamentos]
      summary: Listar todos os pagamentos (paginado)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
          description: Número da página (0-based)
        - in: query
          name: size
          schema: { type: integer, minimum: 1, default: 10 }
          description: Tamanho da página
        - in: query
          name: sort
          schema: { type: string, example: "id,desc" }
          description: Critério de ordenação (`campo,direcao`)
      responses:
        '200':
          description: Lista paginada de pagamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPagamentoResponse'
        '404':
          description: Nenhum pagamento encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/{id}:
    get:
      tags: [Pagamentos]
      summary: Buscar pagamento por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Pagamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagamentoResponse'
        '404':
          description: Pagamento não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
    delete:
      tags: [Pagamentos]
      summary: Exclusão lógica do pagamento (apenas se PENDENTE)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Pagamento inativado com sucesso
        '400':
          description: Regra de negócio violada (não está PENDENTE ou já inativo)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: Pagamento não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/cpf-cnpj/{cpfCnpj}:
    get:
      tags: [Pagamentos]
      summary: Listar pagamentos por CPF/CNPJ
      parameters:
        - in: path
          name: cpfCnpj
          required: true
          schema: { type: string, example: "12345678901" }
      responses:
        '200':
          description: Lista de pagamentos do documento informado
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PagamentoResponse' }
        '404':
          description: Nenhum pagamento encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/codigo/{codigoDebito}:
    get:
      tags: [Pagamentos]
      summary: Listar pagamentos por código de débito
      parameters:
        - in: path
          name: codigoDebito
          required: true
          schema: { type: integer, format: int32, example: 123456 }
      responses:
        '200':
          description: Lista de pagamentos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PagamentoResponse' }
        '404':
          description: Nenhum pagamento encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/status/{status}:
    get:
      tags: [Pagamentos]
      summary: Listar pagamentos por status
      parameters:
        - in: path
          name: status
          required: true
          schema: { $ref: '#/components/schemas/StatusPagamento' }
      responses:
        '200':
          description: Lista de pagamentos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PagamentoResponse' }
        '404':
          description: Nenhum pagamento encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/codigo/{codigoDebito}/cpf-cnpj/{cpfCnpj}:
    get:
      tags: [Pagamentos]
      summary: Buscar pagamento por código de débito e CPF/CNPJ
      parameters:
        - in: path
          name: codigoDebito
          required: true
          schema: { type: integer, format: int32 }
        - in: path
          name: cpfCnpj
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pagamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagamentoResponse'
        '404':
          description: Não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/cpf-cnpj/{cpfCnpj}/status/{status}:
    get:
      tags: [Pagamentos]
      summary: Listar pagamentos por CPF/CNPJ e status
      parameters:
        - in: path
          name: cpfCnpj
          required: true
          schema: { type: string }
        - in: path
          name: status
          required: true
          schema: { $ref: '#/components/schemas/StatusPagamento' }
      responses:
        '200':
          description: Lista de pagamentos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PagamentoResponse' }
        '404':
          description: Nenhum pagamento encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/pagamentos/{id}/status:
    patch:
      tags: [Pagamentos]
      summary: Atualizar status do pagamento
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: novoStatus
          required: true
          schema: { $ref: '#/components/schemas/StatusPagamento' }
          description: |
            Transições válidas:
              - PENDENTE → {PENDENTE, PROCESSADO_COM_SUCESSO, PROCESSADO_COM_FALHA}
              - PROCESSADO_COM_FALHA → PENDENTE
              - PROCESSADO_COM_SUCESSO → não permite alterar
      responses:
        '200':
          description: Status atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagamentoResponse'
        '400':
          description: Regra de negócio violada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: Pagamento não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

components:
  schemas:
    MetodoPagamento:
      type: string
      enum: [BOLETO, PIX, CARTAO_DEBITO, CARTAO_CREDITO]
      example: PIX

    StatusPagamento:
      type: string
      enum: [PENDENTE, PROCESSADO_COM_FALHA, PROCESSADO_COM_SUCESSO]
      example: PENDENTE

    PagamentoRequest:
      type: object
      required: [codigoDebito, cpfCnpj, metodoPagamento, valor]
      properties:
        codigoDebito:
          type: integer
          format: int32
          minimum: 1
          example: 123456
        cpfCnpj:
          type: string
          description: CPF ou CNPJ do pagador (somente dígitos)
          example: "12345678901"
        metodoPagamento:
          $ref: '#/components/schemas/MetodoPagamento'
        numeroCartao:
          type: string
          description: Obrigatório quando metodoPagamento for cartão (crédito/débito)
          maxLength: 20
          example: "5555444433331111"
        valor:
          type: number
          format: double
          minimum: 0.01
          example: 250.00

    PagamentoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoDebito:
          type: integer
          format: int32
          example: 123456
        cpfCnpjPagador:
          type: string
          example: "12345678901"
        metodoPagamento:
          $ref: '#/components/schemas/MetodoPagamento'
        valor:
          type: number
          format: double
          example: 250.00
        status:
          $ref: '#/components/schemas/StatusPagamento'
        ativo:
          type: boolean
          example: true

    PaginatedPagamentoResponse:
      type: object
      description: Estrutura padrão de paginação do Spring Data
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/PagamentoResponse' }
        totalElements:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 5
        size:
          type: integer
          example: 10
        number:
          type: integer
          description: Página atual (0-based)
          example: 0
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false

    Problem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-11-07T12:30:21.123Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Recurso não encontrado"
        message:
          type: string
          example: "Pagamento não encontrado: id=1"
        path:
          type: string
          example: "/api/pagamentos/1"
        validationErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string